// src/generated/models/Product.ts
export interface Product {
    /**
     * 
     * @type {string}
     * @memberof Product
     */
    id: string;
    /**
     * 
     * @type {number}
     * @memberof Product
     */
    score: number;
    /**
     * 
     * @type {string}
     * @memberof Product
     */
    url: string;
    /**
     * 
     * @type {string}
     * @memberof Product
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof Product
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Product
     */
    brand_name: string;
    /**
     * 
     * @type {string}
     * @memberof Product
     */
    image_url: string;
    /**
     * 
     * @type {Price}
     * @memberof Product
     */
    price: Price;
    /**
     * 
     * @type {AvailabilityStatus}
     * @memberof Product
     */
    availability: AvailabilityStatus;
    /**
     * 
     * @type {Array<Variant>}
     * @memberof Product
     */
    variants?: Array<Variant>;
}